Moved wordpress page updater related classes to their own package
