+++ src/main/java/fi/metatavu/soteapi/wordpress/tasks/pages/PageUpdateJob.java
@@ -1,4 +1,4 @@
-package fi.metatavu.soteapi.wordpress.tasks;
+package fi.metatavu.soteapi.wordpress.tasks.pages;
 
 import javax.inject.Inject;
