+++ src/main/java/fi/metatavu/soteapi/wordpress/tasks/pages/PageUpdateTaskModel.java
@@ -1,0 +1,9 @@
+package fi.metatavu.soteapi.wordpress.tasks.pages;
+
+import fi.metatavu.soteapi.wordpress.tasks.AbstractPostUpdateModel;
+
+public class PageUpdateTaskModel extends AbstractPostUpdateModel {
+
+  private static final long serialVersionUID = -7364528550441816286L;
+
+}
